Folder and namespace restructuring
