+++ TollFeeCalculator/CalculateTollFeeService.cs
@@ -1,4 +1,8 @@
-﻿namespace TollFeeCalculator;
+﻿using TollFeeCalculator.Exceptions;
+using TollFeeCalculator.Models;
+using TollFeeCalculator.Policies;
+
+namespace TollFeeCalculator;
 
 public interface ICalculateTollFeeService
 {
