+++ TollFeeCalculator/Models/TollFeeCalculationResult.cs
@@ -1,3 +1,3 @@
-﻿namespace TollFeeCalculator;
+﻿namespace TollFeeCalculator.Models;
 
 public record TollFeeCalculationResult(DateOnly Date, decimal TotalFee, string Currency);
